refactor PPTX generation logic to use pptx_service for improved modularity and maintainability
