+++ reports/07_test_and_refactor.md
@@ -5,7 +5,7 @@
 - **Scope:** Systematic backend cleanup following a full codebase duplication/architecture-debt audit (see `reports/.notes.md` for the complete 9-item backend + 8-item frontend findings list). This report covers the first 5 backend items, tackled one at a time with full verification after each.
 
 ## Change Summary
-- **6 findings resolved**, each verified independently: syntax check, full `pytest` suite (27/27 passing throughout), Docker container reload with no errors, plus targeted behavioral verification for the riskiest changes (exact error-message diffing, live end-to-end test against the real PostgreSQL + pgvector database, mocked OpenAI round-trips).
+- **7 findings resolved**, each verified independently: syntax check, full `pytest` suite (27/27 passing throughout), Docker container reload with no errors, plus targeted behavioral verification for the riskiest changes (exact error-message diffing, live end-to-end test against the real PostgreSQL + pgvector database, mocked OpenAI round-trips, a real generated-and-reopened `.pptx` file).
 - No user-facing API behavior changed except item 1, which makes previously-dead cleanup logic actually run.
 
 ## 1. File Cleanup Worker Wired Up (`999e153`)
@@ -43,9 +43,14 @@
   - `api/v1/ideas.py` is now a thin route module: auth dependency, delegate to the service, `try/except` → `HTTPException(500)` — same error-handling behavior as before, no upgrade in granularity (kept in scope).
 - **Verified:** `POST /api/v1/ideas/topics` without a token still returns `401` (routing intact). With the real `AsyncOpenAI` client mocked out (`unittest.mock.patch`), ran both `ideas_service.generate_topic_ideas()` (JSON-mode response → parsed into `TopicIdea` objects) and `ideas_service.chat_about_topic()` (prompt formatting + message history truncation) end-to-end inside the running container — both produced correct output matching pre-refactor logic exactly.
 
+## 7. Extracted PPTX Export Engine from `presentations.py` God-File
+- **Problem:** `presentations.py` was an 825-line route module that, alongside its route handlers, also contained a full private PPTX-generation engine (`_generate_pptx_from_state`, ~185 lines) plus its SSRF-safe image-fetching helpers (`_fetch_image_bytes_safely`, `_is_public_ip`) and a color-conversion helper (`_hex_to_rgb`) — despite `pptx_service.py` already existing specifically for PPTX concerns (until now handling only extraction/preview conversion).
+- **Fix:** Moved the entire block (`_ALLOWED_IMAGE_HOSTS`, `_is_public_ip`, `_fetch_image_bytes_safely`, `_hex_to_rgb`, and `_generate_pptx_from_state`, renamed to the public `generate_pptx_from_state`) into `pptx_service.py`. The `export-pptx` route now only does auth, DB fetch, delegates to `pptx_service.generate_pptx_from_state(state)` via `run_in_executor`, and formats the HTTP response (the ASCII/UTF-8 `Content-Disposition` filename logic stayed in the route since that's response formatting, not PPTX generation). `presentations.py` dropped from **825 to 568 lines**.
+- **Simplification found along the way:** the moved function used to alias `from pptx import Presentation as PptxPresentation` to avoid colliding with `presentations.py`'s own `Presentation` (the SQLAlchemy model). `pptx_service.py` has no such collision — it already imports the real `pptx.Presentation` at module level for its existing functions — so the alias was dropped in favor of using that import directly.
+- **Verified:** the container stopped cleanly mid-step for an unrelated reason (exit code 0, not a crash from this change) and was restarted. After restart: full `pytest` suite 27/27, `export-pptx` route still returns `401` without a token (routing intact), and — the strongest check — **called `pptx_service.generate_pptx_from_state()` directly with a real 3-slide `PresentationState`** (standard/left/background layouts, real Unsplash image URLs) and re-opened the resulting 234KB output with `python-pptx`: correct slide count, correct titles/bullets, correct sequential slide numbers (1, 2, 3), images embedded successfully (confirming the SSRF-guard helpers still work post-move).
+
 ## Remaining Work (not yet started)
-From the original audit, items 7-9 (backend) and the full frontend list are still open:
-- `presentations.py` is an 825-line god-file — the PPTX export engine belongs in `pptx_service.py`.
+From the original audit, items 8-9 (backend) and the full frontend list are still open:
 - `api/v1/dashboard/planner/planner.py` is the only nested router; should flatten to `api/v1/planner.py`.
 - Unused exception classes (`DatabaseError`, `AuthenticationError`, `ResourceNotFoundError`, `RateLimitError`) are never raised; some "not found" cases incorrectly return 400 via `ValidationError` instead of 404.
 - Frontend: `PresentationEditor.tsx` god-component, duplicated session-storage parsing, duplicated auth-check boilerplate, duplicated auth-page scaffolding, no shared `types/`, no shared error-message util, repeated spinner markup.
@@ -55,4 +60,4 @@ Every item in this report was checked with all of the following before being con
 1. `python -m ast.parse` on every touched file (syntax).
 2. Full backend `pytest` suite — stayed at 27/27 passing after each individual step.
 3. Docker container (`presentation_backend`) log inspection after each change to confirm a clean `--reload` with no import errors or tracebacks.
-4. Item-specific behavioral checks where the risk warranted it: exact exception-message diffing (item 4), live database round-trip with real embedding dimensions (item 5), mocked-OpenAI-client round-trips through the relocated service functions (item 6), and object-identity assertions (`is`) to confirm shared singletons are genuinely shared (items 2, 3).
+4. Item-specific behavioral checks where the risk warranted it: exact exception-message diffing (item 4), live database round-trip with real embedding dimensions (item 5), mocked-OpenAI-client round-trips through the relocated service functions (item 6), a real generated `.pptx` re-opened and inspected slide-by-slide (item 7), and object-identity assertions (`is`) to confirm shared singletons are genuinely shared (items 2, 3).
